엘레베이터 제작 및 배치, 리프트 버튼 배치
버튼 다 배치했음
