+++ Assets/ParfeLatte/AccessEvent.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 733aa7e98aa2f464e8db5a797ebd3a44
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
