+++ Assets/ParfeLatte/AccessEvent/AccessPoint.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eb137dfd61eeaf24481e6db01076bc57
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
