+++ Assets/ParfeLatte/AccessEvent/EventMonster.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c86cd66d8694f934390804fe2c243edd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
