--- Assets/ParfeLatte/KeyCard.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 098612149ae756643a5d4941aeedd896
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
