--- Assets/ParfeLatte/KeyCardEvent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ca00d42fd14216f4aac51d7a3f25a9f5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
