+++ Assets/ParfeLatte/Lift.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 394dc36adf938b84ba7f16d183dc9156
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
