+++ Assets/ParfeLatte/Lift/Elevator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 355e00b488df7564a9040214be050bf9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
