+++ Assets/ParfeLatte/Lift/ElevatorButton.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 06afe3806341a004bb58a5feb2ddf393
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
