+++ Assets/ParfeLatte/Lift/ElevatorCall.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9b418255bfedbd24b871914e9b2b8926
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
