+++ Assets/ParfeLatte/Lift/LiftDownButton.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 25941b0c56f05eb4f98679c411c6cb85
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
