--- Assets/ParfeLatte/LiftTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7465512bf74206f45802fb5506135576
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
