--- Assets/ParfeLatte/LiftUpButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8358f4c81785ee046ae2c356d8b583f6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
