--- Assets/ParfeLatte/OnLift.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0b8f9a7ff6824c24b8ba6503b6fb7dfb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
